requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
pillow==11.0.0
pydantic==2.9.2
pydantic[email]==2.9.2
python-multipart==0.0.16
alembic==1.13.3
deepface==0.0.93
bcrypt==3.2.0
passlib[bcrypt]==1.7.4
tf-keras==2.18.0 
aiomysql==0.2.0
python-jose[cryptography]==3.3.0